backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
python-dotenv==1.0.1
pydantic==2.6.0
aiofiles==23.2.1
python-multipart==0.0.9

# Web scraping
beautifulsoup4==4.12.3

# Image processing
pillow==10.2.0

# Audio processing (for Coqui XTTS)
pydub>=0.25.0
scipy>=1.11.0

# GPU/ML (install separately based on system)
# torch>=2.2.0  # Required for Coqui and SadTalker

# Text Generation - vLLM
# vllm>=0.4.0  # Install separately: pip install vllm

# TTS - Coqui XTTS v2
# TTS>=0.22.0  # Install separately: pip install TTS

# Video - SadTalker
# Install from: https://github.com/OpenTalker/SadTalker
# face-alignment>=1.3.5
# imageio>=2.31.0
# imageio-ffmpeg>=0.4.9

# Optional: Anthropic (for future Claude integration)
# anthropic==0.18.1
